requirements.txt
async-timeout==4.0.2
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
prometheus-client==0.15.0
pyparsing==3.0.9
pyserial==3.5
redis==4.3.5
Werkzeug==2.2.2
